remove imports
Ultimately, imports is the wrong feature. I ended up deciding against
needing to use them to link together all of the ledgers. Instead,
ledgers are included by the workspace config. So the only remaining use
case for imports is bringing accounts into scope. But it makes more
sense to have accounts be global. The only down side I can think of is
it makes navigation without an IDE slightly harder, but it's not that
bad to just grep for "account foo" to find the declaration. I also don't
know how often you'd want to do that anyways.